Pulling back infomation correctly however needs work on algorythm

+++ MandCo.Data.DispatchSLAReport/Models/DisplayData.cs
@@ -4,22 +4,22 @@
     {
         public int TotalOrders {get; set;}
         public int TotalOrdersSLA { get; set; }
-        public float TotalOrdersSLAPct { get; set; }
+        public decimal TotalOrdersSLAPct { get; set; }
 
         public int ExpressOrders { get; set; }
         public int ExpressOrdersSLA { get; set; }
-        public float ExpressOrdersSLAPct { get; set; }
+        public decimal ExpressOrdersSLAPct { get; set; }
 
         public int InternationalOrders { get; set; }
         public int InternationalOrdersSLA { get; set; }
-        public float InternationalOrdersSLAPct { get; set; }
+        public decimal InternationalOrdersSLAPct { get; set; }
 
         public int StandardOrders { get; set; }
         public int StandardOrdersSLA { get; set; }
-        public float StandardOrdersSLAPct { get; set; }
+        public decimal StandardOrdersSLAPct { get; set; }
 
         public int StoreOrders { get; set; }
         public int StoreOrdersSLA { get; set; }
-        public float StoreOrdersSLAPct { get; set; }
+        public decimal StoreOrdersSLAPct { get; set; }
     }
 }
